Enable file size caching by default
Right after a file copying Finder checks size of the written file.
With simple-lock it became possible that VNOP_STRATEGY still in-flight
when getattr is issues to the fuse filesystem.
Because not the whole file is still written to user-space then fuse most likely
will report size of half-copied file. Finder does not like it and crashes
copying.

Enable file size caching to prevent this Finder issue.