Remove inconsistency with novncache flag checking
Either we should add/purge vnodes to cache independently of novncache flag,
or we need to check the flag before every vncache operation. Former option is more widely
used in fuse4x so let's use it everywhere.

Take into account that we still check the flag before vncache *lookup*. It means vncache is
up-to-date but we always bypass during lookup() vnode operation.